pyproject.toml
[project]
name = "data-analyst-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv>=1.1.1,<2.0.0",
    "pandas>=2.3.2,<3.0.0",
    "openai>=1.104.2,<2.0.0",
    "tabulate>=0.9.0,<0.10.0",
    "chainlit (>=2.7.2,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
